fix fonts
