+++ cookowly-frontend/src/pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from 'next/app';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <div className="bg-neutral-900 text-neutral-50 text-content">
+    <div className="bg-neutral-900 text-neutral-50 font-content">
       <Component {...pageProps} />
     </div>
   );
